separated fixtures from test cases and split them into conftest.py by module
